# 10-10EZ | Upgrade performance for introduction page